chore: rename util.ts to intl.ts
